Green pastures for the sheep
Add a single green pastures building that's designed to provide some
food security for the early sheep, before the player will be able to
handle the food production with the sheep.

Handle buildings that can't be built by the player, but only received
in special circumstances (e.g. starting the game with them). This can
be used in the future for the heritage buildings bought with prestige
currency etc.