Recruitment details component
Using the fact that a state with no sheep being selected is a valid
one, display a special component to recruit new sheep to the tribe.
This allows for more information to be displayed than on the button,
as well as minimizes the risk of misclicking and spending a lot of food
by mistake.