Sheep names and starting pack
Add a list with a few sheep names, in order to select one somewhat
randomly when a new sheep is recruited.
Removed the test sheep that the player started with; grant the food
instead for a first sheep.
Fix the price calculation for the new sheep (base and exponent were
mixed up).
Fix the bug when recruiting the first sheep (no ID to add one to).