Fix the storage-respecting production function
Originally, it would reduce the amount over the storage to the amount
of storage, even if it was originally more than the storage capacity
(due to reducing storage capacity in the past, for example). This was
not the intended behavior - we want to keep the amount over the storage
limit in place when new production takes place.