Skeleton of locks and unlocking
For now we use buildings to unlock hunts and some test locks, to keep
it simple.
A `GameObject` class now carries a lock collection, from which locks
can be removed. This is handled by the engine, and objects that can
remove locks have a special interface.
When a last lock is removed, relevant method returns a boolean to denote
 that the object is no longer locked, to facilitate informing the player
 about that, to be implemented in the future.